adding a sleep so nmap isn't pinging like crazy
I ran into one instance where nmap said, "hey port 22 is open. You're
good," which caused the while-loop to end. Then the following ssh
command bombed out. My guess is that i managed to ssh to the server in
the tiny amount of time from between when it opens port 22 and when sshd
begins listening on that port. Hoping this sleep will help alleviate
that from happening again.